storage: move maybeStripInFlightWrites
This patch creates a dedicated file for functions performing server-side
modifications to a batch. The next commit will have a 2nd such function.

Release note: None
