+++ pkg/storage/replica_test.go
@@ -436,101 +436,6 @@ func createReplicaSets(replicaNumbers []roachpb.StoreID) []roachpb.ReplicaDescri
 	return result
 }
 
-// TestMaybeStripInFlightWrites verifies that in-flight writes declared on an
-// EndTxn request are stripped if the corresponding write or query intent is in
-// the same batch as the EndTxn.
-func TestMaybeStripInFlightWrites(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	keyA, keyB, keyC := roachpb.Key("a"), roachpb.Key("b"), roachpb.Key("c")
-	qi1 := &roachpb.QueryIntentRequest{RequestHeader: roachpb.RequestHeader{Key: keyA}}
-	qi1.Txn.Sequence = 1
-	put2 := &roachpb.PutRequest{RequestHeader: roachpb.RequestHeader{Key: keyB}}
-	put2.Sequence = 2
-	put3 := &roachpb.PutRequest{RequestHeader: roachpb.RequestHeader{Key: keyC}}
-	put3.Sequence = 3
-	delRng3 := &roachpb.DeleteRangeRequest{RequestHeader: roachpb.RequestHeader{Key: keyC}}
-	delRng3.Sequence = 3
-	scan3 := &roachpb.ScanRequest{RequestHeader: roachpb.RequestHeader{Key: keyC}}
-	scan3.Sequence = 3
-	et := &roachpb.EndTxnRequest{RequestHeader: roachpb.RequestHeader{Key: keyA}, Commit: true}
-	et.Sequence = 4
-	et.IntentSpans = []roachpb.Span{{Key: keyC}}
-	et.InFlightWrites = []roachpb.SequencedWrite{{Key: keyA, Sequence: 1}, {Key: keyB, Sequence: 2}}
-	testCases := []struct {
-		reqs           []roachpb.Request
-		expIFW         []roachpb.SequencedWrite
-		expIntentSpans []roachpb.Span
-		expErr         string
-	}{
-		{
-			reqs:           []roachpb.Request{et},
-			expIFW:         []roachpb.SequencedWrite{{Key: keyA, Sequence: 1}, {Key: keyB, Sequence: 2}},
-			expIntentSpans: []roachpb.Span{{Key: keyC}},
-		},
-		// QueryIntents aren't stripped from the in-flight writes set on the
-		// slow-path of maybeStripInFlightWrites. This is intentional.
-		{
-			reqs:           []roachpb.Request{qi1, et},
-			expIFW:         []roachpb.SequencedWrite{{Key: keyA, Sequence: 1}, {Key: keyB, Sequence: 2}},
-			expIntentSpans: []roachpb.Span{{Key: keyC}},
-		},
-		{
-			reqs:           []roachpb.Request{put2, et},
-			expIFW:         []roachpb.SequencedWrite{{Key: keyA, Sequence: 1}},
-			expIntentSpans: []roachpb.Span{{Key: keyB}, {Key: keyC}},
-		},
-		{
-			reqs:   []roachpb.Request{put3, et},
-			expErr: "write in batch with EndTxn missing from in-flight writes",
-		},
-		{
-			reqs:           []roachpb.Request{qi1, put2, et},
-			expIFW:         nil,
-			expIntentSpans: []roachpb.Span{{Key: keyA}, {Key: keyB}, {Key: keyC}},
-		},
-		{
-			reqs:           []roachpb.Request{qi1, put2, delRng3, et},
-			expIFW:         nil,
-			expIntentSpans: []roachpb.Span{{Key: keyA}, {Key: keyB}, {Key: keyC}},
-		},
-		{
-			reqs:           []roachpb.Request{qi1, put2, scan3, et},
-			expIFW:         nil,
-			expIntentSpans: []roachpb.Span{{Key: keyA}, {Key: keyB}, {Key: keyC}},
-		},
-		{
-			reqs:           []roachpb.Request{qi1, put2, delRng3, scan3, et},
-			expIFW:         nil,
-			expIntentSpans: []roachpb.Span{{Key: keyA}, {Key: keyB}, {Key: keyC}},
-		},
-	}
-	for _, c := range testCases {
-		var ba roachpb.BatchRequest
-		ba.Add(c.reqs...)
-		t.Run(fmt.Sprint(ba), func(t *testing.T) {
-			resBa, err := maybeStripInFlightWrites(&ba)
-			if c.expErr == "" {
-				if err != nil {
-					t.Errorf("expected no error, got %v", err)
-				}
-				resArgs, _ := resBa.GetArg(roachpb.EndTxn)
-				resEt := resArgs.(*roachpb.EndTxnRequest)
-				if !reflect.DeepEqual(resEt.InFlightWrites, c.expIFW) {
-					t.Errorf("expected in-flight writes %v, got %v", c.expIFW, resEt.InFlightWrites)
-				}
-				if !reflect.DeepEqual(resEt.IntentSpans, c.expIntentSpans) {
-					t.Errorf("expected intent spans %v, got %v", c.expIntentSpans, resEt.IntentSpans)
-				}
-			} else {
-				if !testutils.IsError(err, c.expErr) {
-					t.Errorf("expected error %q, got %v", c.expErr, err)
-				}
-			}
-		})
-	}
-}
-
 // TestIsOnePhaseCommit verifies the circumstances where a
 // transactional batch can be committed as an atomic write.
 func TestIsOnePhaseCommit(t *testing.T) {
